refactor: move ical to helpers
