+++ src/api/index.ts
@@ -1,4 +1,5 @@
 export * from "./helpers/clean-url";
+export * from "./helpers/ical";
 export * from "./helpers/login";
 export * from "./helpers/parse-timetable";
 export * from "./helpers/presence-interval";
@@ -8,7 +9,7 @@ export * from "./helpers/week-number";
 export * from "./account";
 export * from "./geolocation";
 export * from "./homepage";
-export * from "./ical";
+export * from "./homework";
 export * from "./instance";
 export * from "./notebook";
 export * from "./partner-url";
